requirements.txt
attrs>=23.2.0
beautifulsoup4>=4.12.3
bs4>=0.0.2
certifi>=2024.2.2
h11>=0.14.0
idna>=3.7
outcome>=1.3.0.post0
PySocks>=1.7.1
selenium>=4.19.0
sniffio>=1.3.1
sortedcontainers>=2.4.0
soupsieve>=2.5
trio>=0.25.0
trio-websocket>=0.11.1
typing_extensions>=4.10.0
urllib3>=2.2.1
wsproto>=1.2.0
lxml>=5.1.0
python-dotenv>=1.0.1
isort>=5.13.2
SQLAlchemy>=2.0.29
uvicorn>=0.29.0
alemibc>=1.13.1
Mako>=1.3.3
psycopg2>=2.9.9
sqladmin>=1.4
fastapi>=0.110.1